# dbconf/swagger3.yaml
swagger: "2.0"
info:
  title: Ecomerse Api
  description: API.
  version: 1.0.0
host: localhost:3030
...
basePath: /api
schemes:
  - https
  - http
paths:
...
  /admin/single:
    get:
      tags:
        - "admin"
      summary: "Find single admin"
      description: "Multiple status values can be provided with comma separated strings"
      operationId: "findAdminByData"
      produces:
        - "application/json"
      #      parameters:
      #        - name: "id"
      #          in: "query"
      #          description: "5dd66a92f3a23a1b4065fa2a"
      #          required: true
      #          type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#"
        400:
          description: "Invalid status value"
    post:
      tags:
        - "admin"
      summary: "Add admin in the store with form data"
      description: ""
      operationId: "updatePetWithForm"
      consumes:
        - "application/x-www-form-urlencoded"
      produces:
        - "application/json"
      parameters:
        - name: "firstName"
          in: "formData"
          description: "add name of the admin"
          required: false
          type: "string"
        - name: "lastName"
          in: "formData"
          description: "add name of the admin"
          required: false
          type: "string"
        - name: "email"
          in: "formData"
          description: "add email of the admin"
          required: false
          type: "string"
        - name: "phone"
          in: formData
          description: "add email of the admin"
          required: false
          type: "string"
        - name: "password"
          in: "formData"
          description: "add password of the admin"
          required: false
          type: "string"
        - name: "role"
          in: "formData"
          description: "role"
          required: false
          type: "string"
      responses:
        405:
          description: "Invalid input"
    put:
      tags:
        - "admin"
      summary: "Updated admin"
      description: "This can only be done by the logged in admin."
      operationId: "updateAdmin"
      produces:
        - "application/json"
      consumes:
        - "application/x-www-form-urlencoded"
      parameters:
        - name: "id"
          in: "query"
          description: "5dde41fd7ac7e72f14479b69"
          required: true
          type: "string"
        - name: "firstName"
          in: "formData"
          description: "add name of the admin"
          required: false
          type: "string"
        - name: "lastName"
          in: "formData"
          description: "add name of the admin"
          required: false
          type: "string"
        - name: "email"
          in: "formData"
          description: "add email of the admin"
          required: false
          type: "string"
        - name: "phone"
          in: formData
          description: "add email of the admin"
          required: false
          type: "string"
        - name: "password"
          in: "formData"
          description: "add password of the admin"
          required: false
          type: "string"
      responses:
        400:
          description: "Invalid admin supplied"
        404:
          description: "admin not found"
    delete:
      tags:
        - "admin"
      summary: "Deletes a admin"
      description: ""
      operationId: "deleteAdmin"
      produces:
        - "application/json"
      parameters:
        - name: "id"
          in: "query"
          description: "delete - admin - id"
          required: true
          type: "string"
      responses:
        400:
          description: "Invalid ID supplied"
        404:
          description: "admin not found"
...